Files for Spooky Eyes
This is the code and circuit diagram for the Spooky Eyes portion of the haunted house. It includes 5 sets of "eyes" that have strips of LEDs that glow into and out of the darkness, randomly. There is code to include the use of a motion sensor and code for just continuous, random glowing eye action.